backend/requirements.txt
# LangChain Core
langchain
langchain-core
# OpenAI Integration
langchain-openai
openai
# Anthropic Integration
langchain-anthropic
# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai
# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
# Environment Variable Management
python-dotenv
# Machine Learning Utilities
numpy
scikit-learn 

flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-huggingface==0.0.1
langchain-groq==0.0.1
langchain-community==0.0.10
youtube-transcript-api==0.6.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0